Move token authority deployment to colony client
